Moved segments into translation app
