+++ wagtail_localize/translation/utils.py
@@ -10,8 +10,8 @@ from django.db.models import (
     When,
 )
 
-from wagtail_localize.segments import TemplateValue, RelatedObjectValue
-from wagtail_localize.segments.extract import extract_segments
+from wagtail_localize.translation.segments import TemplateValue, RelatedObjectValue
+from wagtail_localize.translation.segments.extract import extract_segments
 
 from .models import (
     Segment,
